upd 0.0.1
git-svn-id: https://fv3.googlecode.com/svn/trunk@71 1faddcec-d5ef-11de-96bf-fbce4fd84a6e

+++ src/fv3/font/ttf/Hhea.java
@@ -50,7 +50,7 @@ public final class Hhea
     public void init(TTFFont font, TTF tables, TTFFontReader reader){
         reader.seek(this.offset+4);
         this.ascent = reader.readUint16();
-        this.descent = reader.readUint16();
+        this.descent = (short)reader.readUint16();
         this.leading = reader.readUint16();
         this.advance = reader.readUint16();
         for (int cc = 0; cc < 11; cc++)
